refactor: remove @fluentui/react-theme-provider (#18710)
* chore: remove @fluentui/react-theme-provider

* Change files

* fix API.md file

* fix deps

* fix tests

* add babel config

* add babel config

* fix useThemeStyleTag usage

* rename perf test

* fix version mismatch
