+++ apps/perf-test/package.json
@@ -19,7 +19,6 @@
     "@fluentui/react-button": "^9.0.0-alpha.52",
     "@fluentui/react-provider": "^9.0.0-alpha.49",
     "@fluentui/react-tabs": "^1.0.0-beta.118",
-    "@fluentui/react-theme-provider": "^9.0.0-alpha.46",
     "@fluentui/react-theme": "^9.0.0-alpha.15",
     "@fluentui/scripts": "^1.0.0",
     "@microsoft/load-themed-styles": "^1.10.26",
