+++ apps/perf-test/src/scenarios/FluentProviderNext.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import { webLightTheme } from '@fluentui/react-theme';
+import { FluentProvider } from '@fluentui/react-provider';
+
+const Scenario = () => <FluentProvider theme={webLightTheme} />;
+
+export default Scenario;
