--- apps/perf-test/src/scenarios/ThemeProviderNext.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-import { webLightTheme } from '@fluentui/react-theme';
-import { ThemeProvider } from '@fluentui/react-theme-provider';
-
-const Scenario = () => <ThemeProvider theme={webLightTheme} />;
-
-export default Scenario;
