+++ apps/vr-tests/package.json
@@ -34,7 +34,6 @@
     "@fluentui/react-provider": "^9.0.0-alpha.49",
     "@fluentui/react-tabs": "^1.0.0-beta.118",
     "@fluentui/react-theme": "^9.0.0-alpha.15",
-    "@fluentui/react-theme-provider": "^9.0.0-alpha.46",
     "@fluentui/scripts": "^1.0.0",
     "@fluentui/storybook": "^1.0.0",
     "@storybook/addons": "6.0.28",
