+++ change/@fluentui-react-accordion-5779e09f-2eb2-44a4-a4b1-7f97d9d69c11.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "remove redundant dependency on react-theme-provider",
+  "packageName": "@fluentui/react-accordion",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
