+++ change/@fluentui-react-components-0b6bd27e-55ed-44bd-a144-27e78b514a17.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "remove @fluentui/react-theme-provider",
+  "packageName": "@fluentui/react-components",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
