+++ change/@fluentui-react-examples-495ebaca-2a07-4d06-a3eb-0b87bf1e3d6a.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "update examples to use @fluentui/react-provider",
+  "packageName": "@fluentui/react-examples",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
