+++ change/@fluentui-react-provider-8a02e661-8c32-4548-b32b-f27a2bad447a.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "move theming capabilities from @fluentui/react-theme-provider",
+  "packageName": "@fluentui/react-provider",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
