+++ change/@fluentui-react-tooltip-241651cf-ded0-4991-8e3f-0ff659ad88ee.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "use react-shared-contexts instead of react-theme-provider",
+  "packageName": "@fluentui/react-tooltip",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
