+++ nx.json
@@ -114,7 +114,6 @@
     "@fluentui/react-tabster": { "tags": ["vNext", "platform:web"], "implicitDependencies": [] },
     "@fluentui/react-text": { "tags": ["vNext", "platform:web"], "implicitDependencies": [] },
     "@fluentui/react-theme": { "implicitDependencies": [] },
-    "@fluentui/react-theme-provider": { "implicitDependencies": [] },
     "@fluentui/react-toggle": { "implicitDependencies": [] },
     "@fluentui/react-tooltip": { "implicitDependencies": [] },
     "@fluentui/react-utilities": { "implicitDependencies": [] },
