+++ packages/react-accordion/package.json
@@ -47,7 +47,6 @@
     "@fluentui/react-make-styles": "^9.0.0-alpha.44",
     "@fluentui/react-tabster": "^9.0.0-alpha.36",
     "@fluentui/react-theme": "^9.0.0-alpha.15",
-    "@fluentui/react-theme-provider": "^9.0.0-alpha.46",
     "@fluentui/react-utilities": "^9.0.0-alpha.29",
     "tslib": "^2.1.0"
   },
