+++ packages/react-components/package.json
@@ -45,7 +45,6 @@
     "@fluentui/react-portal": "^9.0.0-alpha.23",
     "@fluentui/react-provider": "^9.0.0-alpha.49",
     "@fluentui/react-theme": "^9.0.0-alpha.15",
-    "@fluentui/react-theme-provider": "^9.0.0-alpha.46",
     "@fluentui/react-tooltip": "^9.0.0-alpha.49",
     "@fluentui/react-utilities": "^9.0.0-alpha.29",
     "tslib": "^2.1.0"
