+++ packages/react-components/src/index.ts
@@ -2,7 +2,6 @@
 export * from '@fluentui/react-make-styles';
 export * from '@fluentui/react-provider';
 export * from '@fluentui/react-theme';
-export * from '@fluentui/react-theme-provider';
 export { SSRProvider } from '@fluentui/react-utilities';
 
 // Components
