+++ packages/react-examples/package.json
@@ -57,7 +57,6 @@
     "@fluentui/react-slider": "^1.0.0-beta.116",
     "@fluentui/react-tabs": "^1.0.0-beta.118",
     "@fluentui/react-theme": "^9.0.0-alpha.15",
-    "@fluentui/react-theme-provider": "^9.0.0-alpha.46",
     "@fluentui/react-toggle": "^1.0.0-beta.116",
     "@fluentui/react-tooltip": "^9.0.0-alpha.49",
     "@fluentui/react-utilities": "^9.0.0-alpha.29",
