+++ packages/react-provider/src/components/FluentProvider/renderFluentProvider.tsx
@@ -1,22 +1,27 @@
 import * as React from 'react';
 import { FluentProviderState } from './FluentProvider.types';
-import { ProviderContext, TooltipContext } from '@fluentui/react-shared-contexts';
-import { ThemeProvider } from '@fluentui/react-theme-provider';
+import { ProviderContext, TooltipContext, ThemeContext, ThemeClassNameContext } from '@fluentui/react-shared-contexts';
+import { getSlots } from '@fluentui/react-utilities';
 
 /**
  * Render the final JSX of FluentProvider
  */
 export const renderFluentProvider = (state: FluentProviderState) => {
-  const { dir, targetDocument, theme } = state;
+  const { slots, slotProps } = getSlots(state, []);
+  const { dir, targetDocument } = state;
 
   // eslint-disable-next-line react-hooks/rules-of-hooks
   const value = React.useMemo(() => ({ dir, targetDocument }), [dir, targetDocument]);
 
   return (
-    <ThemeProvider {...state} theme={theme} targetDocument={state.targetDocument}>
-      <ProviderContext.Provider value={value}>
-        <TooltipContext.Provider value={state.tooltipContext}>{state.children}</TooltipContext.Provider>
-      </ProviderContext.Provider>
-    </ThemeProvider>
+    <ProviderContext.Provider value={value}>
+      <ThemeContext.Provider value={state.theme}>
+        <ThemeClassNameContext.Provider value={state.className ?? ''}>
+          <TooltipContext.Provider value={state.tooltipContext}>
+            <slots.root {...slotProps.root}>{state.children}</slots.root>
+          </TooltipContext.Provider>
+        </ThemeClassNameContext.Provider>
+      </ThemeContext.Provider>
+    </ProviderContext.Provider>
   );
 };
