+++ packages/react-provider/src/components/FluentProvider/useFluentProviderStyles.ts
@@ -1,5 +1,5 @@
 import { makeStyles, mergeClasses } from '@fluentui/react-make-styles';
-import { ThemeProviderState } from './ThemeProvider.types';
+import { FluentProviderState } from './FluentProvider.types';
 
 const useStyles = makeStyles({
   root: theme => ({
@@ -12,7 +12,7 @@ const useStyles = makeStyles({
 });
 
 /** Applies style classnames to slots */
-export const useThemeProviderStyles = (state: ThemeProviderState) => {
+export const useFluentProviderStyles = (state: FluentProviderState) => {
   // Theme override is passed here to use a proper theme otherwise it will usa a theme from a parent
   const styles = useStyles();
   state.className = mergeClasses(styles.root, state.className);
