--- packages/react-theme-provider/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["plugin:@fluentui/eslint-plugin/react"],
-  "root": true
-}
