--- packages/react-theme-provider/.npmignore
@@ -1,34 +1,0 @@
-*.api.json
-*.config.js
-*.log
-*.nuspec
-*.test.*
-*.yml
-.editorconfig
-.eslintrc*
-.eslintcache
-.gitattributes
-.gitignore
-.vscode
-coverage
-dist/storybook
-dist/*.stats.html
-dist/*.stats.json
-dist/demo
-fabric-test*
-gulpfile.js
-images
-index.html
-jsconfig.json
-node_modules
-results
-src/**/*
-!src/**/examples/*.tsx
-!src/**/docs/**/*.md
-!src/**/*.types.ts
-temp
-tsconfig.json
-tsd.json
-tslint.json
-typings
-visualtests
