--- packages/react-theme-provider/config/api-extractor.json
@@ -1,3 +1,0 @@
-{
-  "extends": "@fluentui/scripts/api-extractor/api-extractor.common.json"
-}
