--- packages/react-theme-provider/config/tests.js
@@ -1,7 +1,0 @@
-/** Jest test setup file. */
-
-const { configure } = require('enzyme');
-const Adapter = require('enzyme-adapter-react-16');
-
-// Configure enzyme.
-configure({ adapter: new Adapter() });
