--- packages/react-theme-provider/just.config.ts
@@ -1,3 +1,0 @@
-import { preset } from '@fluentui/scripts';
-
-preset();
