--- packages/react-theme-provider/package.json
@@ -1,64 +1,0 @@
-{
-  "name": "@fluentui/react-theme-provider",
-  "version": "9.0.0-alpha.46",
-  "description": "Fluent UI React theme provider component, hook, and theme related utilities.",
-  "main": "lib-commonjs/index.js",
-  "module": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "sideEffects": false,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/fluentui"
-  },
-  "license": "MIT",
-  "scripts": {
-    "build": "just-scripts build",
-    "bundle": "just-scripts bundle",
-    "clean": "just-scripts clean",
-    "code-style": "just-scripts code-style",
-    "just": "just-scripts",
-    "lint": "just-scripts lint",
-    "start": "just-scripts dev:storybook",
-    "start-test": "just-scripts jest-watch",
-    "test": "just-scripts test",
-    "update-snapshots": "just-scripts jest -u"
-  },
-  "devDependencies": {
-    "@fluentui/babel-make-styles": "^9.0.0-alpha.19",
-    "@fluentui/eslint-plugin": "^1.3.1",
-    "@fluentui/jest-serializer-make-styles": "^9.0.0-alpha.20",
-    "@fluentui/react-conformance": "^0.4.1",
-    "@fluentui/scripts": "^1.0.0",
-    "@types/classnames": "^2.2.9",
-    "@types/enzyme": "3.10.3",
-    "@types/enzyme-adapter-react-16": "1.0.3",
-    "@types/jest": "24.9.1",
-    "@types/react": "16.9.42",
-    "@types/react-dom": "16.9.10",
-    "@types/react-test-renderer": "^16.0.0",
-    "enzyme": "~3.10.0",
-    "enzyme-adapter-react-16": "^1.15.0",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-test-renderer": "^16.3.0"
-  },
-  "dependencies": {
-    "@fluentui/react-make-styles": "^9.0.0-alpha.44",
-    "@fluentui/react-shared-contexts": "^9.0.0-alpha.15",
-    "@fluentui/react-theme": "^9.0.0-alpha.15",
-    "@fluentui/react-utilities": "^9.0.0-alpha.29",
-    "tslib": "^2.1.0"
-  },
-  "peerDependencies": {
-    "@types/react": ">=16.8.0 <18.0.0",
-    "react": ">=16.8.0 <18.0.0"
-  },
-  "beachball": {
-    "tag": "alpha",
-    "disallowedChangeTypes": [
-      "major",
-      "minor",
-      "patch"
-    ]
-  }
-}
