--- packages/react-theme-provider/src/ThemeProvider.ts
@@ -1,1 +1,0 @@
-export * from './components/ThemeProvider/index';
