--- packages/react-theme-provider/src/common/isConformant.ts
@@ -1,13 +1,0 @@
-import { isConformant as baseIsConformant, IsConformantOptions } from '@fluentui/react-conformance';
-
-export function isConformant<TProps = {}>(
-  testInfo: Omit<IsConformantOptions<TProps>, 'componentPath'> & { componentPath?: string },
-) {
-  const defaultOptions: Partial<IsConformantOptions<TProps>> = {
-    disabledTests: ['has-docblock'],
-    asPropHandlesRef: true,
-    componentPath: module!.parent!.filename.replace('.test', ''),
-  };
-
-  baseIsConformant(defaultOptions, testInfo);
-}
