--- packages/react-theme-provider/src/components/ThemeProvider/index.ts
@@ -1,4 +1,0 @@
-export * from './ThemeProvider';
-export * from './ThemeProvider.types';
-export * from './useThemeProvider';
-export * from './renderThemeProvider';
