--- packages/react-theme-provider/src/index.ts
@@ -1,3 +1,0 @@
-export * from './ThemeProvider';
-
-export { useTheme } from '@fluentui/react-shared-contexts';
