--- packages/react-theme-provider/webpack.config.js
@@ -1,5 +1,0 @@
-const resources = require('../../scripts/webpack/webpack-resources');
-
-module.exports = resources.createBundleConfig({
-  output: 'FluentUIReactThemeProvider',
-});
